dashboard/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
stripe==11.4.0
python-multipart==0.0.20
